Some edits to work with GCC 6.1 on Mac OS X: - Added -Wno-deprecated-declarations because of auto_ptr - Explicitely defined set_config for uint64_t - Manually cast uint64_t as size_t in ostream - Explicitely code min() for differing types - Added missing headers

+++ src/output/output_format.cpp
@@ -120,7 +120,7 @@ void XML_format::print_query_epilog(const DAA_query_record &r, Text_buffer &out)
 		<< "  <Iteration_stat>" << '\n'
 		<< "    <Statistics>" << '\n'
 		<< "      <Statistics_db-num>" << ref_header.sequences << "</Statistics_db-num>" << '\n'
-		<< "      <Statistics_db-len>" << config.db_size << "</Statistics_db-len>" << '\n'
+		<< "      <Statistics_db-len>" << (size_t) config.db_size << "</Statistics_db-len>" << '\n'
 		<< "      <Statistics_hsp-len>0</Statistics_hsp-len>" << '\n'
 		<< "      <Statistics_eff-space>0</Statistics_eff-space>" << '\n'
 		<< "      <Statistics_kappa>").print_d(score_matrix.k()) << "</Statistics_kappa>" << '\n'
